+++ libs/core/src/state/settings/by_monitor.rs
@@ -3,9 +3,12 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use ts_rs::TS;
 
-use crate::{resource::WidgetId, state::by_widget::ThirdPartyWidgetSettings};
+use crate::{
+    resource::WidgetId,
+    state::{by_widget::ThirdPartyWidgetSettings, WorkspaceId},
+};
 
-#[derive(Clone, Debug, Default, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema, TS)]
 pub struct MonitorSettingsByWidget(HashMap<WidgetId, ThirdPartyWidgetSettings>);
 
 impl MonitorSettingsByWidget {
@@ -16,9 +19,22 @@ impl MonitorSettingsByWidget {
     }
 }
 
-#[derive(Clone, Debug, Default, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(default, rename_all = "camelCase")]
 pub struct MonitorConfiguration {
     /// dictionary of settings by widget
     pub by_widget: MonitorSettingsByWidget,
+    /// Id of the wallpaper collection to use in this monitor.\
+    /// If not set, the default wallpaper collection will be used.
+    pub wallpaper_collection: Option<uuid::Uuid>,
+    /// dictionary of settings by workspace on this monitor
+    pub by_workspace: HashMap<WorkspaceId, WorkspaceConfiguration>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema, TS)]
+#[serde(default, rename_all = "camelCase")]
+pub struct WorkspaceConfiguration {
+    /// Id of the wallpaper collection to use in this workspace.\
+    /// If not set, the monitor's wallpaper collection will be used.
+    pub wallpaper_collection: Option<uuid::Uuid>,
 }