+++ src/ui/react/wallpaper_manager/app.tsx
@@ -1,11 +1,11 @@
-import { getCurrentWindow } from "@tauri-apps/api/window";
 import { useEffect } from "react";
 
 import { MonitorContainers } from "./modules/Monitor/infra.tsx";
+import { Widget } from "@seelen-ui/lib";
 
 export function App() {
   useEffect(() => {
-    getCurrentWindow().show();
+    Widget.getCurrent().ready();
   }, []);
 
   return <MonitorContainers />;